requirement.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.3
djangorestframework==3.13.1
h11==0.13.0
html5lib==1.1
idna==3.3
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
Unidecode==1.3.3
urllib3==1.26.8
webencodings==0.5.1
wsproto==1.1.0
